refactor: ViewController 클래스 복원 및 패키지 변경
