+++ src/main/java/roomescape/controller/rest/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.rest;
 
 import java.net.URI;
 import java.util.List;
