+++ src/main/java/roomescape/controller/rest/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.request;
+package roomescape.controller.rest.request;
 
 import java.time.LocalDate;
