+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import roomescape.controller.request.ReservationRequest;
+import roomescape.controller.rest.request.ReservationRequest;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
