+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 
-import roomescape.controller.request.ReservationRequest;
+import roomescape.controller.rest.request.ReservationRequest;
 import roomescape.exception.EntityExistsException;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
